Split api document main.ts into multiple files.
